// apps/backend/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Updates {
  blocked: Generated<boolean>;
  created_at: Generated<Timestamp | null>;
  description: string | null;
  update_id: Generated<string>;
  username: string;
}

export interface DB {
  updates: Updates;
}